# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 2D RPG MMO API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{characterId}/inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get character inventory
      security:
        - bearerAuth: []
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"

    post:
      operationId: addInventoryItem
      tags: [inventory]
      summary: Add an item to inventory
      security:
        - bearerAuth: []
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddInventoryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /characters/{characterId}/inventory/{itemId}:
    delete:
      operationId: removeInventoryItem
      tags: [inventory]
      summary: Remove an item from inventory
      security:
        - bearerAuth: []
      parameters:
        - name: characterId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        characterId:
          type: integer
        itemKey:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [weapon, armor, potion, material, quest]
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        quantity:
          type: integer
        attackBonus:
          type: integer
        defenseBonus:
          type: integer
        hpBonus:
          type: integer
        description:
          type: string
        equipped:
          type: boolean
        acquiredAt:
          type: string
          format: date-time
      required: [id, characterId, itemKey, name, type, rarity, quantity, attackBonus, defenseBonus, hpBonus, description, equipped, acquiredAt]
    AddInventoryItemBody:
      type: object
      properties:
        itemKey:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [weapon, armor, potion, material, quest]
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        quantity:
          type: integer
        attackBonus:
          type: integer
        defenseBonus:
          type: integer
        hpBonus:
          type: integer
        description:
          type: string
      required: [itemKey, name, type, rarity]
...