# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 2D RPG MMO API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new account
 */
export const registerBodyUsernameMin = 3;
export const registerBodyUsernameMax = 20;

export const registerBodyPasswordMin = 6;

export const registerBodyCharacterNameMin = 2;
export const registerBodyCharacterNameMax = 24;

export const RegisterBody = zod.object({
  username: zod
    .string()
    .min(registerBodyUsernameMin)
    .max(registerBodyUsernameMax),
  password: zod.string().min(registerBodyPasswordMin),
  characterName: zod
    .string()
    .min(registerBodyCharacterNameMin)
    .max(registerBodyCharacterNameMax),
  characterClass: zod.enum(["warrior", "mage", "archer", "rogue"]),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  character: zod.object({
    id: zod.number(),
    userId: zod.number(),
    name: zod.string(),
    class: zod.enum(["warrior", "mage", "archer", "rogue"]),
    level: zod.number(),
    xp: zod.number(),
    xpToNextLevel: zod.number(),
    hp: zod.number(),
    maxHp: zod.number(),
    mp: zod.number(),
    maxMp: zod.number(),
    attack: zod.number(),
    defense: zod.number(),
    speed: zod.number(),
    posX: zod.number(),
    posY: zod.number(),
    mapId: zod.string(),
    gold: zod.number(),
    kills: zod.number(),
    deaths: zod.number(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user and character
 */
export const GetMeResponse = zod.object({
  token: zod.string(),
  character: zod.object({
    id: zod.number(),
    userId: zod.number(),
    name: zod.string(),
    class: zod.enum(["warrior", "mage", "archer", "rogue"]),
    level: zod.number(),
    xp: zod.number(),
    xpToNextLevel: zod.number(),
    hp: zod.number(),
    maxHp: zod.number(),
    mp: zod.number(),
    maxMp: zod.number(),
    attack: zod.number(),
    defense: zod.number(),
    speed: zod.number(),
    posX: zod.number(),
    posY: zod.number(),
    mapId: zod.string(),
    gold: zod.number(),
    kills: zod.number(),
    deaths: zod.number(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get a character by ID
 */
export const GetCharacterParams = zod.object({
  characterId: zod.coerce.number(),
});

export const GetCharacterResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  class: zod.enum(["warrior", "mage", "archer", "rogue"]),
  level: zod.number(),
  xp: zod.number(),
  xpToNextLevel: zod.number(),
  hp: zod.number(),
  maxHp: zod.number(),
  mp: zod.number(),
  maxMp: zod.number(),
  attack: zod.number(),
  defense: zod.number(),
  speed: zod.number(),
  posX: zod.number(),
  posY: zod.number(),
  mapId: zod.string(),
  gold: zod.number(),
  kills: zod.number(),
  deaths: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update character stats (equip item, change appearance)
 */
export const UpdateCharacterParams = zod.object({
  characterId: zod.coerce.number(),
});

export const UpdateCharacterBody = zod.object({
  equippedWeaponId: zod.number().nullish(),
  equippedArmorId: zod.number().nullish(),
});

export const UpdateCharacterResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  class: zod.enum(["warrior", "mage", "archer", "rogue"]),
  level: zod.number(),
  xp: zod.number(),
  xpToNextLevel: zod.number(),
  hp: zod.number(),
  maxHp: zod.number(),
  mp: zod.number(),
  maxMp: zod.number(),
  attack: zod.number(),
  defense: zod.number(),
  speed: zod.number(),
  posX: zod.number(),
  posY: zod.number(),
  mapId: zod.string(),
  gold: zod.number(),
  kills: zod.number(),
  deaths: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get character inventory
 */
export const GetInventoryParams = zod.object({
  characterId: zod.coerce.number(),
});

export const GetInventoryResponseItem = zod.object({
  id: zod.number(),
  characterId: zod.number(),
  itemKey: zod.string(),
  name: zod.string(),
  type: zod.enum(["weapon", "armor", "potion", "material", "quest"]),
  rarity: zod.enum(["common", "uncommon", "rare", "epic", "legendary"]),
  quantity: zod.number(),
  attackBonus: zod.number(),
  defenseBonus: zod.number(),
  hpBonus: zod.number(),
  description: zod.string(),
  equipped: zod.boolean(),
  acquiredAt: zod.coerce.date(),
});
export const GetInventoryResponse = zod.array(GetInventoryResponseItem);

/**
 * @summary Add an item to inventory
 */
export const AddInventoryItemParams = zod.object({
  characterId: zod.coerce.number(),
});

export const AddInventoryItemBody = zod.object({
  itemKey: zod.string(),
  name: zod.string(),
  type: zod.enum(["weapon", "armor", "potion", "material", "quest"]),
  rarity: zod.enum(["common", "uncommon", "rare", "epic", "legendary"]),
  quantity: zod.number().optional(),
  attackBonus: zod.number().optional(),
  defenseBonus: zod.number().optional(),
  hpBonus: zod.number().optional(),
  description: zod.string().optional(),
});

/**
 * @summary Remove an item from inventory
 */
export const RemoveInventoryItemParams = zod.object({
  characterId: zod.coerce.number(),
  itemId: zod.coerce.number(),
});

/**
 * @summary Get top players by level and XP
 */
export const getLeaderboardQueryLimitDefault = 20;

export const GetLeaderboardQueryParams = zod.object({
  limit: zod.coerce.number().default(getLeaderboardQueryLimitDefault),
});

export const GetLeaderboardResponseItem = zod.object({
  rank: zod.number(),
  characterId: zod.number(),
  name: zod.string(),
  class: zod.string(),
  level: zod.number(),
  xp: zod.number(),
  kills: zod.number(),
  gold: zod.number(),
});
export const GetLeaderboardResponse = zod.array(GetLeaderboardResponseItem);

/**
 * @summary Get count and summary of online players
 */
export const GetOnlinePlayersResponse = zod.object({
  count: zod.number(),
  players: zod.array(
    zod.object({
      characterId: zod.number(),
      name: zod.string(),
      level: zod.number(),
      class: zod.string(),
      mapId: zod.string(),
    }),
  ),
});